Final Summary Stats output
The summary stats filter for genes DE for both up and down regulation for each sample. the descriptive stats indicate that each of these genes, based on the read counts, were pre-disposed to being up-regulated or down-regulated because of their difference between conditions. it is hypothesised that these values are not being acconted for int he normalisation and therefore these genes keep being identified as DE up and down despite the metadata not identifying them as such